Add proxy_utils module and refactor proxy_helpers
