+++ helpers/proxy_helpers.py
@@ -1,13 +1,1 @@
-import bpy
-
-
-def enable_proxy():
-    """Enable proxies if possible."""
-    if bpy.ops.clip.proxy_on.poll():
-        bpy.ops.clip.proxy_on()
-
-
-def disable_proxy():
-    """Disable proxies if possible."""
-    if bpy.ops.clip.proxy_off.poll():
-        bpy.ops.clip.proxy_off()
+from .proxy_utils import enable_proxy, disable_proxy
