+++ helpers/proxy_utils.py
@@ -1,0 +1,13 @@
+import bpy
+
+
+def enable_proxy():
+    """Enable proxies if possible."""
+    if bpy.ops.clip.proxy_on.poll():
+        bpy.ops.clip.proxy_on()
+
+
+def disable_proxy():
+    """Disable proxies if possible."""
+    if bpy.ops.clip.proxy_off.poll():
+        bpy.ops.clip.proxy_off()
